extracted similar methods to a new parent class (Contact.java) to minimise reuse of code everywhere
